# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gym Management Chatbot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface SessionInput {
  sessionId: string;
  /** @nullable */
  phone?: string | null;
}

export interface ChatSession {
  sessionId: string;
  /** @nullable */
  botMode?: string | null;
  /** @nullable */
  mode?: string | null;
  /** @nullable */
  pendingStep?: string | null;
  /** @nullable */
  pendingData?: string | null;
  updatedAt: string;
  isNew: boolean;
}

export interface ChatMessageInput {
  sessionId: string;
  message: string;
  /** @nullable */
  phone?: string | null;
}

export interface ChatButton {
  label: string;
  value: string;
}

export interface ChatMessage {
  id: string;
  role: string;
  content: string;
  buttons?: ChatButton[];
  timestamp: string;
}

export interface ChatResponse {
  messages: ChatMessage[];
  session: ChatSession;
}

export interface Member {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  gender: string;
  age: number;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  height?: number | null;
  /** @nullable */
  goal?: string | null;
  plan: string;
  joiningDate: string;
  expiryDate: string;
  paymentStatus: string;
  /** @nullable */
  trainerId?: number | null;
  /** @nullable */
  trainerName?: string | null;
  /** @nullable */
  bmi?: number | null;
  status: string;
  createdAt: string;
}

export interface MemberInput {
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  gender: string;
  age: number;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  height?: number | null;
  /** @nullable */
  goal?: string | null;
  plan: string;
  joiningDate: string;
  paymentStatus: string;
  /** @nullable */
  trainerId?: number | null;
}

export interface MemberUpdate {
  name?: string;
  phone?: string;
  /** @nullable */
  email?: string | null;
  gender?: string;
  age?: number;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  height?: number | null;
  /** @nullable */
  goal?: string | null;
  plan?: string;
  paymentStatus?: string;
  /** @nullable */
  trainerId?: number | null;
  status?: string;
}

export interface RenewalInput {
  plan: string;
  paymentStatus: string;
}

export interface ProgressEntry {
  id: number;
  memberId: number;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  bmi?: number | null;
  /** @nullable */
  notes?: string | null;
  recordedAt: string;
}

export interface ProgressEntryInput {
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface AttendanceRecord {
  id: number;
  memberId: number;
  /** @nullable */
  memberName?: string | null;
  /** @nullable */
  checkIn?: string | null;
  /** @nullable */
  checkOut?: string | null;
  date: string;
}

export type AttendanceInputType = typeof AttendanceInputType[keyof typeof AttendanceInputType];


export const AttendanceInputType = {
  check_in: 'check_in',
  check_out: 'check_out',
} as const;

export interface AttendanceInput {
  memberId: number;
  type: AttendanceInputType;
  /** @nullable */
  phone?: string | null;
}

export type AttendanceStatsByDayItem = {
  day: string;
  count: number;
};

export interface AttendanceStats {
  todayCount: number;
  weekCount: number;
  monthCount: number;
  byDay: AttendanceStatsByDayItem[];
}

export interface Trainer {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  specialization?: string | null;
  /** @nullable */
  bio?: string | null;
  memberCount: number;
  createdAt: string;
}

export interface TrainerInput {
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  specialization?: string | null;
  /** @nullable */
  bio?: string | null;
}

export interface TrainerUpdate {
  name?: string;
  phone?: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  specialization?: string | null;
  /** @nullable */
  bio?: string | null;
}

export interface TrainerBooking {
  id: number;
  trainerId: number;
  /** @nullable */
  trainerName?: string | null;
  memberId: number;
  /** @nullable */
  memberName?: string | null;
  scheduledAt: string;
  /** @nullable */
  notes?: string | null;
  status: string;
  createdAt: string;
}

export interface TrainerBookingInput {
  trainerId: number;
  memberId: number;
  scheduledAt: string;
  /** @nullable */
  notes?: string | null;
}

export interface DietPlan {
  id: number;
  title: string;
  goal: string;
  content: string;
  /** @nullable */
  trainerId?: number | null;
  /** @nullable */
  trainerName?: string | null;
  createdAt: string;
}

export interface DietPlanInput {
  title: string;
  goal: string;
  content: string;
  /** @nullable */
  trainerId?: number | null;
}

export interface WorkoutPlan {
  id: number;
  title: string;
  goal: string;
  content: string;
  /** @nullable */
  trainerId?: number | null;
  /** @nullable */
  trainerName?: string | null;
  createdAt: string;
}

export interface WorkoutPlanInput {
  title: string;
  goal: string;
  content: string;
  /** @nullable */
  trainerId?: number | null;
}

export interface Lead {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  interest?: string | null;
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface LeadInput {
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  interest?: string | null;
}

export interface LeadUpdate {
  status?: string;
  /** @nullable */
  notes?: string | null;
}

export interface Broadcast {
  id: number;
  title: string;
  message: string;
  type: string;
  targetAudience: string;
  sentCount: number;
  createdAt: string;
}

export interface BroadcastInput {
  title: string;
  message: string;
  type: string;
  targetAudience: string;
}

export interface DashboardSummary {
  totalMembers: number;
  activeMembers: number;
  expiredMembers: number;
  todayAttendance: number;
  totalTrainers: number;
  totalLeads: number;
  monthlyRevenue: number;
  expiringThisMonth: number;
}

export interface ActivityItem {
  id: string;
  type: string;
  description: string;
  timestamp: string;
}

export type ListMembersParams = {
search?: string;
status?: string;
plan?: string;
};

export type ListAttendanceParams = {
memberId?: number;
date?: string;
};

export type ListDietPlansParams = {
goal?: string;
trainerId?: number;
};

export type ListWorkoutPlansParams = {
goal?: string;
};

export type ListLeadsParams = {
status?: string;
};