# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gym Management Chatbot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chat/session:
    post:
      operationId: createChatSession
      tags: [chat]
      summary: Create or restore a chat session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionInput"
      responses:
        "200":
          description: Session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSession"
  /chat/message:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message and get bot response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: Bot response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
  /chat/session/{sessionId}:
    delete:
      operationId: resetChatSession
      tags: [chat]
      summary: Reset a chat session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── MEMBERS ───────────────────────────────────────────────────────────────
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Leads list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Submit an inquiry / lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Lead created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update lead status / notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"

  # ── BROADCAST ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # ── CHAT SCHEMAS ────────────────────────────────────────────────────────
    SessionInput:
      type: object
      properties:
        sessionId:
          type: string
        phone:
          type: ["string", "null"]
      required:
        - sessionId
    ChatSession:
      type: object
      properties:
        sessionId:
          type: string
        botMode:
          type: ["string", "null"]
        mode:
          type: ["string", "null"]
        pendingStep:
          type: ["string", "null"]
        pendingData:
          type: ["string", "null"]
        updatedAt:
          type: string
        isNew:
          type: boolean
      required:
        - sessionId
        - isNew
        - updatedAt
    ChatMessageInput:
      type: object
      properties:
        sessionId:
          type: string
        message:
          type: string
        phone:
          type: ["string", "null"]
      required:
        - sessionId
        - message
    ChatMessage:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
        content:
          type: string
        buttons:
          type: array
          items:
            $ref: "#/components/schemas/ChatButton"
        timestamp:
          type: string
      required:
        - id
        - role
        - content
        - timestamp
    ChatButton:
      type: object
      properties:
        label:
          type: string
        value:
          type: string
      required:
        - label
        - value
    ChatResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
        session:
          $ref: "#/components/schemas/ChatSession"
      required:
        - messages
        - session

    # ── MEMBER SCHEMAS ──────────────────────────────────────────────────────
...
    Lead:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        interest:
          type: ["string", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - phone
        - status
        - createdAt
    LeadInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        interest:
          type: ["string", "null"]
      required:
        - name
        - phone
    LeadUpdate:
      type: object
      properties:
        status:
          type: string
        notes:
          type: ["string", "null"]

    # ── BROADCAST SCHEMAS ───────────────────────────────────────────────────
...