# lib/api-zod/src/generated/types/chatMessageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gym Management Chatbot API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatMessageInput {
  sessionId: string;
  message: string;
  /** @nullable */
  phone?: string | null;
}