# lib/api-zod/src/generated/types/sessionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gym Management Chatbot API
 * OpenAPI spec version: 0.1.0
 */

export interface SessionInput {
  sessionId: string;
  /** @nullable */
  phone?: string | null;
}